[update]版本号3.0.0，ldrobotsensor SDK3.0 framework import success.

+++ clean_build.sh
@@ -1,6 +1,19 @@
 #!/bin/bash
 #Author: David Hu
-#Date: 2021-11
-rm -rf build/
-echo "del ./build/"
+#Date: 2022-09
+if [ -e "./build" ];then
+  rm -rf build/
+  echo "del ./build/"
+fi
+
+if [ -e "./install" ];then
+  rm -rf install/
+  echo "del ./install/"
+fi
+
+if [ -e "./log" ];then
+  rm -rf log/
+  echo "del ./log/"
+fi
+
 echo "del is ok....."
